Got ranges into both V0 and V1
